last commit

+++ app/Http/Controllers/ListingController.php
@@ -35,6 +35,7 @@ class ListingController extends Controller
         if($request->hasFile('logo')){
             $formFields['logo'] = $request->file('logo')->store('logos','public');
         }
+        $formFields[ 'user_id' ]= auth()->id();
         // dd($formFields);
         Listing::create($formFields);
         return redirect('/')->with('message','Listing created successfully!');
@@ -44,6 +45,11 @@ class ListingController extends Controller
         return view('listings.edit',['listing'=> $listing]);
     }
     public function update(Request $request ,Listing $listing){
+
+        if($listing->user_id!=auth()->id()){
+            return abort(403, "Unauthorized action");
+        };
+
         $formFields= $request->validate([
             'title'=>'required',
             'company'=>'required',
@@ -62,7 +68,13 @@ class ListingController extends Controller
         return back()->with('message','Listing updated successfully!');
     }
     public function destroy(Listing $listing){
+        if($listing->user_id!=auth()->id()){
+            return abort(403, "Unauthorized action");
+        };
         $listing->delete();
         return redirect('/')->with('message','Listing Deleted Successfully');
+    }
+    public function manage(){
+        return view('listings.manage',['listings'=>auth()->user()->listings()->get()]);
     }
 }
