+++ routes/web.php
@@ -17,27 +17,30 @@ use App\Http\Controllers\ListingController;
 |
 */
 
-Route::get('/', [ListingController::class,'index']);
+Route::get('/', [ListingController::class,'index'])->middleware('auth');
 
-Route::get('/listings/create',[ListingController::class,'create']);
+Route::get('/listings/create',[ListingController::class,'create'])->middleware('auth');
 
-Route::post('/listings',[ListingController::class, 'store']);
+Route::post('/listings',[ListingController::class, 'store'])->middleware('auth');
 
-Route::get('/listings/{listing}/edit',[ListingController::class,'edit']);
+Route::get('/listings/{listing}/edit',[ListingController::class,'edit'])->middleware('auth');
 
-Route::put('/listings/{listing}',[ListingController::class,'update']);
+Route::put('/listings/{listing}',[ListingController::class,'update'])->middleware('auth');
 
-Route::delete('/listings/{listing}',[ListingController::class,'destroy']);
+Route::delete('/listings/{listing}',[ListingController::class,'destroy'])->middleware('auth');
 
-Route::get('/register',[UserController::class,'create']);
+Route::get('listings/manage',[ListingController::class,'manage'])->middleware('auth');
+
+Route::get('/register',[UserController::class,'create'])->middleware('guest');
 
 Route::post('/users',[UserController::class,'store']);
 
-Route::post('/logout',[UserController::class,'logout']);
+Route::post('/logout',[UserController::class,'logout'])->middleware('auth');
 
-Route::get('/login',[UserController::class,'login']);
+Route::get('/login',[UserController::class,'login'])->name('login')->middleware('guest');
 
 Route::post('/users/authenticate',[UserController::class,'authenticate']);
+
 
 
 Route::get('/listings/{listing}',[ListingController::class,'show']);
